+++ epf/Reprocessor.cpp
@@ -37,7 +37,7 @@ Reprocessor::Reprocessor(const VoxelKey& k, int numPoints, int pointSize,
     //  =>
     // log2(numPoints / MaxPointsPerNode) = 2n
 
-    m_levels = std::ceil(log2((double)numPoints / MaxPointsPerNode) / 2);
+    m_levels = (int)std::ceil(log2((double)numPoints / MaxPointsPerNode) / 2);
 
     // We're going to steal points from the leaf nodes for sampling, so unless the
     // spatial distribution is really off, this should be fine and pretty conservative.
